requirements.txt
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.60.0
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
